# Optimize `RemoteOutputChecker.addOutputToDownload`