--- a/src/main/java/com/google/devtools/build/lib/remote/RemoteOutputChecker.java
+++ b/src/main/java/com/google/devtools/build/lib/remote/RemoteOutputChecker.java
@@ -37,10 +37,10 @@ import com.google.devtools.build.lib.analysis.configuredtargets.RuleConfiguredTa
 import com.google.devtools.build.lib.analysis.test.TestProvider;
 import com.google.devtools.build.lib.clock.Clock;
 import com.google.devtools.build.lib.remote.options.RemoteOutputsMode;
-import com.google.devtools.build.lib.remote.util.ConcurrentPathTrie;
 import com.google.devtools.build.lib.skyframe.SkyFunctions;
 import com.google.devtools.build.lib.vfs.PathFragment;
 import com.google.devtools.build.skyframe.MemoizingEvaluator;
+import java.util.concurrent.ConcurrentSkipListSet;
 import java.util.function.Predicate;
 import java.util.function.Supplier;
 import javax.annotation.Nullable;
@@ -65,7 +65,7 @@ public class RemoteOutputChecker implements OutputChecker {
 
   @Nullable private Clock clock;
 
-  private final ConcurrentPathTrie pathsToDownload = new ConcurrentPathTrie();
+  private final ConcurrentArtifactPathTrie pathsToDownload = new ConcurrentArtifactPathTrie();
 
   public RemoteOutputChecker(
       String commandName,
@@ -246,12 +246,8 @@ public class RemoteOutputChecker implements OutputChecker {
     }
   }
 
-  public void addOutputToDownload(ActionInput file) {
-    if (file instanceof Artifact && ((Artifact) file).isTreeArtifact()) {
-      pathsToDownload.addPrefix(file.getExecPath());
-    } else {
-      pathsToDownload.add(file.getExecPath());
-    }
+  private void addOutputToDownload(ActionInput file) {
+    pathsToDownload.add(file);
   }
 
   private boolean shouldAddTopLevelTarget(@Nullable ConfiguredTarget configuredTarget) {
@@ -361,6 +357,37 @@ public class RemoteOutputChecker implements OutputChecker {
             return functionName.equals(SkyFunctions.TARGET_COMPLETION)
                 || functionName.equals(SkyFunctions.ASPECT_COMPLETION);
           });
+    }
+  }
+
+  /**
+   * A specialized concurrent trie that stores paths of artifacts and allows checking whether a
+   * given path is contained in (in the case of a tree artifact) or exactly matches (in any other
+   * case) an artifact in the trie.
+   */
+  private static final class ConcurrentArtifactPathTrie {
+    // Invariant: no path in this set is a prefix of another path.
+    private final ConcurrentSkipListSet<PathFragment> paths = new ConcurrentSkipListSet<>();
+
+    /**
+     * Adds the given {@link ActionInput} to the trie.
+     *
+     * <p>The caller must ensure that no object's path passed to this method is a prefix of any
+     * previously added object's path. Bazel enforces this for non-aggregate artifacts. Callers must
+     * not pass in {@link TreeFileArtifact}s (which have exec paths that have their parent tree
+     * artifact's exec path as a prefix) or non-Artifact {@link ActionInput}s that violate this
+     * invariant.
+     */
+    void add(ActionInput input) {
+      paths.add(input.getExecPath());
+    }
+
+    /** Checks whether the given {@link PathFragment} is contained in an artifact in the trie. */
+    boolean contains(PathFragment execPath) {
+      // By the invariant of this set, if a prefix of execPath is present, it must sort right before
+      // it (or be equal to it).
+      var floorPath = paths.floor(execPath);
+      return floorPath != null && execPath.startsWith(floorPath);
     }
   }
 }